feat: レイアウトとヘッダーコンポーネントを更新 - メタデータをサイトの定数に基づいて設定し、ナビゲーションをヘッダーに統合。不要なナビゲーションコンポーネントを削除し、テストを追加。
