--- src/components/layouts/Navigation/index.ts
@@ -1,1 +1,0 @@
-export * from './Navigation'
